Bug#11748945 MAKE KILL RELIABLE (MAIN.KILL FAILS RANDOMLY)
Analysis:
For OSX, POLLHUP event is not implemented for poll(). As a 
result, shutdown() / close() system call does not break
poll(). As client end of the socket is closed,
connection is broken from the client side whereas server 
side service socket is kept open as poll() waits for 28800 
secs (default interactive_timeout and read_timeout value). 

Sleep calls in kill.test are changed to debug sync under
rev-2875.109.79. It fixed all the sporiadic issues related 
to kill connection. This issue is regression of WL#4896 
found only in OSX. There are no failures found on pb2 
for platforms except OSX.

Fix:
Use select() instead of poll() only for OSX as select() 
breaks when close() system call is invoked on the service
socket.