Implement mysql-js ession.save and batch.save
Save delegates to UserContext.save

Save is similar to persist, taking a domain object to store in the database.
The difference is that if there is already a row in the database, the
non-pk values in the object will update the corresponding non-pk values
in the database.

Test case:
  save objects; verify that they were stored
  batch save objects; verify that they were stored
  save an object; save the same key but different values; verify the new values
  persist even objects in the database; save all objects and verify that all were stored

TODO: save with not all fields in the domain object; this will require analysis 
of the domain object and creation of a different save call with only the fields present