Fix for bug#14709490: EXECUTION OF 'PUSHED JOINS' MAY EXHAUST JOB BUFFERS
A pushed scan query, with multiple 'bushy' lookup children could flood
the job buffers by producing multiple LQHKEYREQ's for each row
returned from the scan. Normally we allows a 1::4 fanout between
consumed and produced signals, but this is not sufficient for
such a pushed join: There could be an (almost) unlimited number of rows
returned from the scan (and multiple such executing in parallel)
so we could overflow the jobbuffers by allowing any fanout > 1.

This fix introduce sequentialization points in the SPJ execution:
Any lookup children which otherwise would have been executed in parallel
with another lookup sibling is tagged with T_EXEC_SEQUENTIAL.
When such a TreeNode attempt to send a LQHKEYREQ, the operation is
deferred (queued). It is later resumed when its parallel child
completes (CONF or REF) one of its own REQuests.

Thereby we have restricted the fanout for bushy scan-lookup joins
to 1::1.

Furthermore, it changes the 'pushed query planer' (>= V7.2) in the
handler interface such that we now try to avoid some of the bushines
among lookup childrens in the pushed query. 

Due to the sequentialization this patch enforce among siblings, it
is misleading to have an EXPLAINed query plan which show that siblings
are executed in parallel. By also making them sequentially in the
pushed join, we may take advantage of the selectivity of previously
equi-joined children, such that non-matching branches in the query
Tree are not executed.

There are also the refactorications:

 1) Introduce Dbspj::cleanupBatch() which is intended to do all
    required resorce dealloc and required reinits in preparation for
    a new batch of result rows.

    Also removes 'reset' argument from releaseRequestBuffers and instead
    do required reset from callee (::cleanupBatch()) if required.

2)  Moved Dbspj::registerActiveCursor to a place
    closer to where it is normally used, and closer to other functional
    dependent methods.