BUG#12395512: NULL BYTE IN GROUP_CONCAT()'S SEPARATOR CLAUSE               BREAKS MYSQLDUMP
Analysis:
---------
When the view definition contains special character
in the SEPARATOR clause of the GROUP_CONCAT aggregate
function, the 'mysqldump' utility creates invalid
view definition.

Hence the reload of such a dump file results in syntax
error and the view is not re-created.

While creating the view definition, character escaping is
not performed on the string specified in the SEPARATOR
clause. This results in the creation of the invalid view
definition.

Fix:
---
Performed character escaping on the string specified in the
SEPARATOR clause of the GROUP_CONCAT aggregate function
while creating the view definition.

Other issues fixed by this patch:
--------------------------------
a) The string literal for SEPARATOR clause contains incorrect
value when the charset defined by 'character_set_client' and
'character_set_connection' are different.
b) The string literal for SEPARATOR clause contains incorrect
value in I_S.VIEWS table when the 'character_set_client' is
different from system charset(UTF8).