Bug #15866623 ADD CHECKS FOR #SQL-IB NAME CLASH IN ALTER TABLE
In an ALTER TABLE where InnoDB rebuilds the table, we generate two
temporary table names: one for the copy of the original table, and
another that the original table will be renamed to when the tables are
ready to be switched.

prepare_inplace_alter_table_dict(): If a table with the duplicate name
exists in the data dictionary cache or in the persistent data
dictionary, refuse the operation. Previously, we would attempt to load
the table definition, triggering an assertion failure if the table
existed in the dictionary cache. Do not attempt to drop the table with
the duplicate name, in order not to spoil any manual data recovery.

row_merge_drop_table(): Remove the parameter nonatomic, which will
always be passed as nonatomic=false.

rb:1549 approved by Jimmy Yang