BUG#14801541:TRANSACTION IN ENGINE WITHOUT 2PC SUPPORT IS LOGGED AS XID-ENDED
Problem: Transaction in engine without two phase
commit support is logged as XID event instead of
Query event. This is a problem since the event is
supposed to be restored while recovery, but the Xid
event will be skipped.

Background: This bug has come into picture after a
refactoring which moves code from binlog_commit() to
MYSQL_BIN_LOG::commit() and binlog group commit implementation.

Fix: Restored the previous behaviour for the engines
without two-phase commit support. This bug doesnot affect
any of the engines in the server, so the testing has to be
done in mysql-cluster with NDB