Bug#15831502 Assert failed in Item_outer_ref::fix_after_pullout on subquery
When investigating calls to Item::fix_after_pullout() some time ago,
we did not find any uses of Item_outer_ref, thus we added an assert for
this case. Now, we had such a case, and hence also a crash.

Code was revisited, and we found that we could simply call
Item_ref::fix_after_pullout() from this function, as long as
*ref == outer_ref.