BUG#11747931: SHOW CREATE VIEW OUTPUT DEPENDS ON DEFINER GRANTS
Analysis:
--------

SHOW CREATE VIEW output varies depending on the user
account specified in the DEFINER clause if:
a) The user account does not exist.
b) The user account exists but does not hold SELECT and
   SHOW VIEW privileges.
The VIEW definition output consists of the schema name
under the conditions mentioned above though the view is
defined in the same schema where the view tables reside.

Code which is responsible for how identifiers are printed
in view definitions resides in Item_ident::print() method.
To figure out the way in which the identifier should be
printed, this method uses the 'compact_view_format' member of
Item_ident::cached_table->belong_to_view. The
'compact_view_format' member is set only during the SHOW
CREATE VIEW operation. It is set when the view tables and
the view reside in the same schema. If it is not set then
the the schema name is appended to the view definition.

The 'cached_table' was not populated under the error scenario
where the account in the DEFINER clause was non-existent or
did not hold the required privileges. Hence the schema name
was appended to the VIEW definition.

Fix:
---
A new query type 'QT_COMPACT_FORMAT' is introduced to determine
whether to append the schema name in the VIEW definition of
SHOW CREATE VIEW output. Also 'TABLE_LIST::compact_view_format'
was removed as part of this fix.