Bug#14787508 - ASSERT TRX INTR || BUF_FLUSH_GET_DIRTY_PAGES_COUNT(ID) == 0 FLUSH TABLES FOR EXP
There was a race in the purge stop code. If thread T1 changed the state to
PURGE_STATE_STOP and before purge was suspended (purge_sys->running == false)
if another thread T2 also tried to stop purge it would see that the purge
sys state was stopped and it would proceed. The fix is for T2 to wait for
the purge_sys->running to be set to false.

Approved by Jimmy Yang over IM.