Bug#15867985 	STRESS TESTS CAUSES DBUG SERVER TO EXIT WITH SIGNAL 6
When run with --mysqld=--binlog-order-commits=1 (default) mysqld executes a 
piece of BGC logics that tames a group of threads in a way that they give up
their commit execution path to a leader thread.
This preemption appeared to have a specific to DBUG_ON build flaw in that
the leader and a follower thread could modify a thread-specific area
concurrently which led to a segfault.

Fixed with refining the preemption logics to be thread-safe for the DBUG_ON build.