Fix for Bug#15907122 INCORRECT HANDLING OF JOB-BUFFERS ALMOST FULL - 'SLEEPLOOP 10'
This fix introduce a 'thr_wait*' (Sort of a condition/mutex)
which is used to wait for job buffers to become available when
we are in an 'job buffers almost full' situation. This 'thr_wait'
will be signaled when a job buffer has been consumed & freed.

This fix is a collection of the following 5 sub patches:

******

Fix mixed up comment & naming wrt. whether 'm_recv_transporters' and
'm_has_data_transporters' is a bitmask of NodeId or Tranporters.
(It *is* NodeId bitmasks)

******
Add dump of job buffer utilization before we 
crash due to 'job buffer full', and to 'sleeploop 10'
reports where we have to wait due to almost full job buffers.

******
Change the signature of the yield() function such that it can take a more
general '*check_callback()' function as arguments.

Change the existing functions currently used as 'callbacks' 
to the new (relaxed) signature.
******
Refactor: Splitt out compute_free_buffers_in_queue() 
from compute_max_signals_to_execute().

New function contains common code intended for reuse in later patches.
******
update_sched_config() will 'yield' the CPU and wait for more job buffers
to become available when it is about to run out of job buffers.

The yield() call will wait on a 'thr_wait' object, which may be sent
a 'wakeup()' when the waiting condition has been resolved (by another thread)

However, update_schec_config() used the 'thr_wait' object intended to be used 
to wait for more *incomming signal* - What we actually have to wait for 
are signals to be *consumed* by the destination thread. Luckily there is 
also defined a 'max wait' of 1ms which currently will wakeup the thread .

This patch introduce a own 'thr_wait' object which which is signaled by the
consumer, and we now wait on this object.

Furthermore the patch also avoid a situation where a thread could end up
waiting for itself.