Bug #14845445 : access allowed for invalid hashes
Added a check for invalid native password and old password
plugin hashes when reading the system tables.
Since the function is reused in GRANT and SET PASSWORD 
added some assert()s that it will not fail the check there.

Fixed a spotted problem with the optimization of the plugin name
comparisons when adding users through CREATE USER.

Added test cases.