BUG#15847447 - THD->TRANSACTION.STMT.IS_EMPTY()                || (THD->STATE_FLAGS & OPEN_TABLES_STATE::BACKU
DESCRIPTION:
When a view is based on a Heap (In memory) backed table and this
Table is altered to drop some field which was in use by the view, then
a SELECT on this view results in an ASSETION failure when the server
is started with Bin-Log enabled.


ANALYSIS :
The triggered assert in open_normal_and_derived_tables() checks that
the current statement transaction is empty if something fails during
opening of tables and handling of derived tables. This makes sure that
there is nothing that the statement has done that needs to be rolled
back at this point.

The reason that the assert is triggered in this case is that a
statement transaction is started by THD::binlog_query() during
open_tables() -> open_table_entry_fini() if we are opening a
HEAP table for the first time, subsequently mysql_handle_derived()
fails while processing the view since a column has disappeared. 

FIX :
During implicit_emptied processing in the call to THD::binlog_query() direct
is set to TRUE, as a result the event is immediately written to the log bypassing
a transaction, hence the assert is not triggered.