Bug#15908291 UT_AD(0) ASSERTION FAILURE IN ALTER TABLE...ADD PRIMARY KEY
row_merge_tuple_cmp(): Remove a bogus assertion. We can get a
duplicate entry if dup=NULL, in case we are executing ADD PRIMARY KEY,
ADD [UNIQUE] INDEX. The duplicate entry may not yet have been detected
in the clustered index (PRIMARY KEY), and we will get a duplicate in
the secondary index (or in a unique secondary index if any of the
unique key columns are NULL). We will have to ignore the duplicate in
the secondary index and let the sort on the primary key to detect it
later.

rb:1610 approved by Sunny Bains