Bug#15883071 - WINDOWS: ASSERT DUE TO INCORRECTLY AUTOSIZED INNODB_BUFFER_POOL_INSTANCES
The changes for WL#6216 (Changes to Server defaults in 5.6) introduced
a Windows-only calculation for the number of buffer pool instances when
the buffer pool size became greater than 1.3Gb. This is needed on Windows
32-bit systems because it was found that they have trouble allocating
blocks of memory to mysqld.exe in chunks greater than 1.3 Gb. 

This calculation set the number of instances by dividing the total buffer
pool size by 128Mb, which was determined to be an acceptable buffer pool
instance size.  

Even though this calculation is only needed for 32-bit systems it used
the macro _WIN32 which is also defined for 64-bit Windows.  This extra
adjustment to the number of buffer pool instances is not needed on
64-bit Windows systems.  So this patch uses
   #if defined(__WIN__) && !defined(_WIN64)
to enclose the extra 32-bit Windows calculation.

In addition, it uses ut_min(MAX_BUFFER_POOLS, ...) in case that number
is ever changed to something less than what is allowed on a 32-bit system
(4G buffer pool/128M per pool=32 instances, currently, MAX_BUFFER_POOLS==64)

Approved by Sunny Bains at http://rb.no.oracle.com/rb/r/1579/