Accepted new result files for ndb_join_pushdown_*.test.
The optimizer has become better in removing redundant 
where-predicate terms which are also covered by a (eq-)ref.
This cause some 'Using where' and '... pushed condition...' to
be removed from the EXPLAIN output

NOTE: There is also a single change in query result of one 
of the test due to this change. It is questionable whether
the new or old result is correct 
(due to MySQL infamous implicit type conversion)

Olav Sandstaa is invest this, anyway this diff is not 
Cluster related.