BUG#15833516 - SQL_SLAVE_SKIP_COUNTER DOES NOT SKIP ERRORS WITH GTID_MODE ON
=== Problem ===

'SET GLOBAL sql_slave_skip_counter = N' skips the next N events
from the master. This is useful for recovering from replication
stops caused by a statement. But this works only with a server
not running in GTID mode. When a server is configured to run in
GTID mode, empty transactions have to be executed to do the same.

'sql_slave_skip_counter' should not be used with GTIDs. It will
not work the way a user would expect it to. In particular, after
a failover or reconnect, the skipped transaction will be
re-executed again, which is consistent and unavoidable with GTID
logic, but clearly not what the user wanted.

The correct way to skip transactions when GTIDs are enabled is
to execute an empty transaction, i.e.: 
1. Get the GTID of the failing transaction;
2. Execute SET GTID_NEXT = '<gtid of failing transaction>'; 
   (Note that the server skips transactions if GTID_NEXT is set
   to a GTID that already exists in GTID_DONE.)
3. Execute COMMIT; 

This will make the transaction be skipped permanently.

=== Fix ===

The bug is fixed by generating an error message when a user
tries to set sql_slave_skip_counter when gtid_mode = ON.
1) Created a macro "ER_SQL_SLAVE_SKIP_COUNTER_NOT_SETTABLE_IN_GTID_MODE".
2) Added a check under check_slave_skip_counter() to check if GTID mode
   is set. If yes, an error is generated at the client.