Bug#11753843:BINARY AND RELAY LOG FILE NAMES DEPEND ON PID FILE NAME
Problem:
=======
When Binary and relay log names are not specified by user,
the default binary log .NNNNNN and .index files are created
in the data directory using the PID file as basename. The 
same is true for the relay log file, slow log file and query
log file.

According to the documentation, the default names should be
based on the hostname.

Analysis:
========
"rpl_make_log_name" is the function which constructs the 
basenames for binary and relay log files.   This function
takes three arguments, first the "basename" supplied by
user through option, second the "pid-file" name and third
the extension to the file name. The second argument
"pid-file" is used as basename in the absence of user 
specified basename value.

Fix:
===
In "rpl_make_log_name" function call second parameter i.e
"pid-file" is being replaced with "default_logfile_name".
"default_logfile_name" holds the hostname. While testing
the above fix another problem was observed. When user
specifies the option like "slow_query_log_file=" or
"general_log_file=" in the conf file without specifying
the file name it causes an invalid memory free. The
code snippet is given below.

if (!VAR || !*VAR)
{
    my_free(VAR); /* it could be an allocated empty string "" */
    VAR= ALT;
}

Memory for all options are allocated by using "malloc"
within a memroot. When user specifies "empty" string for
"slowlog" and "querylog" files the above code tries to
free the memory for the empty string and reassign the pointer.
Freeing a part of memory within a memroot is incorrect this
causes the invalid memory free error. Since it is an empty
string and memroots are automatically cleared upon exit
the above "my_free" call is not required. Hence removed.