BUG#14059662 - DDL ON EVENT SCHEDULER MEMORY LEAK WITH                SKIP-GRANT-TABLES ENABLED
BACKGROUND:
When MySQL server is started with --skip-grant-tables option
Event Scheduler is disabled,i.e the event scheduler is rendered
nonoperational.
At low level, Event_queue and Event_scheduler object are not
instantiated.
When an event DDL like Create Event or Alter Event is performed 
during this time, it is added in mysql.event table and appropriate
operations either create or update are done respectively.
Then memory is allocated for the event queue element to be 
created, But it is not put in the event queue since event queue
object has not been instantiated in the beginning itself.

While doing server shutdown, when it cleans up scheduler's resources
The memory for the Event Queue Element which is created is not 
deallocated, destructor for it is not called and it hogs the memory 
without deleting the element. Valgrind reported memory leak 
because of this issue.

FIX:
As a fix for this bug, A Check has been added in Create and 
Update event code. If the Event Scheduler is in DISABLED state,
Memory for the event queue element is not allocated.
We are allowed to create the event even when the scheduler is 
DISABLED.
Later When the server is restarted again without skip-grant-tables,
it picks up the event from mysql.event table and starts executing the 
event normally.