Avoid duplicated TcConnectionRec's in LQHs hashlist of such objects.
Internaly in each LQH we keep a hash structure of TcConnectionrec
which are identified by:
  - Id of Transaction owning this TcConnectionrec.
  - A 'tcOpRec' id which uniquely(*below) identify this TcConnectionRec
    within this specific transaction.
  - An optional 'hashHi' id used for SCANREQs in cases where 'tcOpRec'
    on its own cant provide uniqueness.
    This is required in cases where there are multiple (internal) clients
    producing REQs where the uniqueness is only guaranteed within
    each client. Currently the only such client is the SPJ block.

The internal clients of NDB does *not* guarantee hash uniqueness
for LQHKEYREQs as described above (SPJ, node restart ..). However,
these requests are all 'long', 'dirtyOp'-requests and thus never
required to be searched after in the hash table.

This fix exploits this property of such 'dirtyOp' and thus never
inserts its TcConnectionRec in the hash lists. Duplicates
in the hash list should then be avoided.