Bug#15855723 MYSQLDUMP OPTION TO IGNORE 'SHOW CREATE TABLE ..'              ERROR
Added a new option --ignore-error to ignore specific errors.
This option would accept a user supplied comma-separated list
of error numbers which later is matched with the errors that
occur during the course of execution of mysqldump. This option
however would be superseded by the use of --force option, in
which case all the errors will be ignored.

Added a testcase.