Bug#15849654: Valgrind failure in JOIN::compare_costs_of_subquery_strategies
This function reads an accumulated cost figure from the position object
of the table that the subquery evaluation is attached to.
Normally, the cost figures have already been calculated, but for a
loosescan table, the position object is a local object which is then
copied into position, without having a value for accumulated cost figures
set.

This is fixed by explicitly updating the accumulated cost figure in this case.