Bug#14237992: Hook race condition in Binlog Group Commit causing failures on Semi-Sync
==Problem==
When executing, semi sync depends on two hooks to know which
events need a reply from the slave. These are the after_flush and
before_send_event hooks, being the first one responsible for
adding the event to the hash of events that end one of the active
transactions (and because of this need a reply) and the second
one checks if the event is in that hash.

In a typical run: 

1) In a separated thread the mysql_binlog_send is sending events
  and when it ends its loop it checks for more events in the log
  holding the log_lock. 
2) The mysql_binlog_send waits on a condition for more events.
3) During the commit a new query writes to the binlog holding the
  log lock releasing it when finished.
4) The after_flush hook is executed and the end event
  is added to the hash.
5) A signal is emitted and the mysql_binlog_send thread awakes
  and proceeds to send the event to the slave executing the
  before_send_event hook that signals that a reply is needed from
  the slave.

The problem is that right before the step 4), the step 1) can
happen, and in this case the mysql_binlog_send will not stop in
the condition but will continue execution because a new event is
already in the log. In this execution the before_send_event hook
can be executed before the after_flush resulting in a false
return value. Being instructed that this event doesn't need a
reply, the slave will never send it causing a time out.

==Fix== 
The after_flush should be executed under the log_lock and to
accomplish this the hook was moved to be right after the binlog
flush. This also means a change to the hook interface as the
synced flag no longer makes sense stepping up the replication
interface version to version 0x0200. This change also implies an
behavior change in case of an error during the synchronization
process. As it is today, if a sync problem occurs the hook will
never be executed and semi sync will eventually time out and be
turned off when the event is sent to the slave. This no longer
happens, the hook is now always executed meaning no semi sync
time out will exist.