Bug#14760210: ALTER TABLE OF PARTITIONED TABLE IS NON-ATOMIC
The problem: After the first partition has successfully committed its
changes (in commit_inplace_alter_table()) all the other partitions
must be successfully completed.
Currently if there is a crash, then the committed partitions would have
the new definition and the others the old definition.
If the server is shutting down, then the other partitions would not
succeed in committing, resulting in the same state.
If there is a duplicate key error in the other partitions, the result
would be the same (could be possible to fix during error handling,
but still affected by crash/shutdown).

When partitions are out-of-sync there can be crashes/asserts in
underlying handlers.

To avoid crashing, we disable inplace alter for partitioned
tables.

Approved as rb 1645.