Bug #15844519 "USER-DEFINED PARTITIONING AND FOREIGN KEYS IN CLUSTER".
It was impossible to create NDB table with user defined partitioning
and a foreign key using CREATE TABLE statement. OTOH this limitation
was not enforced consistently, so it was possible to create NDB table
with a foreign key and then add partitioning to it using ALTER TABLE, 
thus creating a table which was impossible to backup/restore through 
mysqldump.

The issue was caused by a check in code which prohibited creation of
tables with foreign keys on partitioned tables unless they were
auto-partitioned. Since storage engines with native partitioning,
e.g. NDB, can support both user-defined partitions and foreign keys 
this limitation turned out to be too strict for them.

This patch changes the code to prohibit creation of partitioned 
tables with foreign keys only for storage engines which don't 
support partitioning natively (i.e. require use of ha_partition 
storage engine).

Patch comes without a test case as the only engine that supports
partitioning natively is NDB, which is not available in the main 
branch of 5.6. So this patch was tested manually by applying fix 
to the cluster tree and running tests there.