Bug#15951568 FAILED ASSERT IN DIGEST_COPY
For some specially crafted queries, the following assert
in digest_copy() can fail:

  if (from->m_byte_count > 0)
  {
    ...
  }
  else
  {
    DBUG_ASSERT(! from->m_full); /* can fail */
    DBUG_ASSERT(from->m_byte_count == 0);
    ...
  }

The issue is that a digest storage can be both 'full',
in a sense that adding content to it caused overflow,
and yet 'empty' with a byte count of 0,
in the sense that no content was effectively ever added.

This happens when an overflow occurs for the first query token.

This assert is over optimistic, and the value of m_full is of no consequence here.

The fix is to simply remove this assert.