Bug #13545447: RPL_ROTATE_LOGS FAILS DUE TO CONCURRENCY ISSUES IN REP. CODE
Currently the assert tests in the rpl_rotate_logs test fail to
give a proper error when a failure happens. This is due to the
fact that the debug routines invoked by rpl_assert.inc expect
some info that is usually set by rpl_init.inc

The rotate_logs test now has extra connections and information
about the number of connected servers so that when printing the
debug messages the routines in show_rpl_debug_info can access
both the master and server data.