Bug#14798445: XML FUNCTIONS: XPATH NOT WORKING WITH STORED               FUNCTIONS (REGRESSION FROM 5.5)
Item_func_xml_func::fix_length_and_dec() requires the second
argument to an XML function to be constant.
After WL#4443 a stored program cannot be considered
constant until after the tables are locked, and locking
is no longer done in resolve phase.

Fix: Change Item_func_xml_func::fix_length_and_dec()
to verify that it will at least be constant during execution.
Defer the parsing of the second argument to the execution
phase if it is not constant in resolve phase.

Details: 
The basic issue is that the stored procedure implementation
does not check the validity of declarations with
DETERMINISTIC or NO SQL.  Hence, there is no easy way to
check whether locking of tables are actually needed before
executing a stored procedure.  (WL#6725 has been created for
this task.)

Since we cannot know whether a stored procedure will
actually need to lock tables, we will have to delay all
stored procedure execution until after tables have been
locked.  Hence, if an XPATH expression contains stored
procedures, we will delay parsing of this expression until
execution phase.

The direct reason for the regression, as discussed in the
bug report, is that Item::const_item is no longer a static
property for expressions involving stored procedure since it
has been changed to depend on tables_locked_cache.  (WL#6726,
has been created for this issue.)  Furthermore,
since constness property is cached for Item_func objects, it
will not change when tables are later locked.

Hence, we cannot use const_item() in order to verify that an
XPATH expression is const.  (As shown by the bug, it will
give false positives.)  Instead we will use
const_during_execution().  This will be true for expression
containing stored procedures only if all stored procedures
is declared as DETERMINISTIC.  (Otherwise, the
RAND_TABLE_BIT of used_tables() will be set.)

Should a user declare a non-deterministic function as
deterministic, it may slip through the
const_during_execution() "filter".  However, since
const_item() will be false, XPATH parsing will be delayed
until execution.  At this point necessary tables will be
locked, so we should not risk any crash etc.  However, query
results may not be correct, but users are warned in the
manual about unspecified behavior if DETERMINISTIC is not
used correctly.

Parsing of XPATH expressions before this patch always
happened in fix_fields(), which is run on every execution of
a statement.  Hence, Items created during parsing was
already created in non-permanent MEMROOT even when prepared
statements are involved.  Hence, moving allocation to
execution phase should not change anything in this respect.