bug#14798275 GLOBAL GTID VARIABLES RELEASED TO EARLY DURING SHUTDOWN
Problem: NDB engine encounters an error when embedded server shutdown,
if binlog is opened. It happens because binlog related things are closed
before NDB engine is closed, but the things(e.g GTID) may need when NDB
engine is closing.

Fix: Handler provides a hook ha_binlog_end which tells engines to finish
binlog related works before server closes binlog. But it was not compiled
in embedded server. In this patch it is compiled in embedded server, just
like a common mysql server.

No way to test the patch on mysql server without NDB. But there is a test
ndb_basic can verify this patch on ndb branches.