Bug#14639605 - FTS: FAILING ASSERTION: RESULT != FTS_INVALID
rb://1673 Approved by Jimmy

When a transaction start, the trx->fts_trx is NULL. If the first
operation of this transaction is set savepoint, since trx->fts_trx
is NULL, the savepoint info will not be set for fts. Then, when
rollback to savepoint issues, we will do nothing for fts.
So, we need to copy the savepoints to the trx->fts_trx when it was
created.