Bug#13992148 - IMPROPER EXECUTION OF QUERY WITH "IN (...UNION...)"
Problem:
A query with IN clause and have UNION operator on left, 	
executes improperly.


Analysis:-

In case of IN clause, we check whether the number of column on the left and 
right expression of IN are same or not.

After that we are fixing the field(prepare the subselect engine)of left 
expression of IN clause.

In case of UNION operator we got the number of column after we prepare 
subselect_union_engine.Hence we dont have the correct value of number of 
column, so DEBUG_ASSERT got failed.


Possible Solution:-
First we check fix the field of left side expression and then check the number 
of column.