Bug#15953730 NDB : SLAVE_ALLOW_BATCHING BROKEN IN 7.2
The fix for this bug was merged up from mysql-5.5-cluster-7.2 to 
mysql-5.6-cluster-7.3.

However, in 7.2, this was an MCP patch on Server code.
In 7.3, there is no MCP patch, the fix is upstream in the server code.

So rather than fix as an MCP patch in 7.3, we'd prefer to fix this in
the upstream (5.6) server code, and merge down.

For this reason, the merged fix in 7.3 is backed-out (here), but the
testcase (ndb_rpl_batch) is left in.

This testcase will remain red until a fix can be merged-down from 
a 5.6 tree.