BUG#15971607: BINLOG IS DISABLED WHEN SERVER IS STARTED WITH --SKIP-INNODB
When server was started with both --log-bin and --skip-innodb, binlog 
was incorrectly disabled, what caused that no events were logged at all.
This happened because binlog was used only when two handlertons were 
registered. This didn't take in account that we can only have one 
handlerton, the case where InnoDB is disabled and binlog enabled.

Fixed logging initialization to also consider the case where InnoDB is
disabled and binlog enabled.