Bug#15913195 - FIL_MAKE_IBBACKUP_OLD_NAME FUNCTION CAN CAUSE BUFFER OVERFLOW
Fixed the buffer overflow by passing the right offset of the 
buffer to ut_sprintf_timestamp_without_extra_chars()

No testcase because this is fixed and tested in meb
branches(mysql-5.5-meb and
mysql-5.6-meb). It is possible only to test in meb branches because of
the
UNIV_HOTBACKUP flag

Approved by Marko. rb#1620