Bug#4789249 - LOCKING ASSERT IN FK UPDATE CASCADE CODE, LOCK->TRX->LOCK.WAIT_LOCK == LOCK Bug#14707091 SEGV IN LOCK_DEADLOCK_LOCK_PRINT(), NULL LOCK PTR
Joining T1 does a deadlock check due to wait 

T1 is selected as victim (but not rolled back), for any number of reasons.

We try and resolve more deadlocks 
  - T2 is selected as a victim and rolled back 
  - This grants T1 the lock and is no longer the victim.

Since T1 was originally selected as a victim, when we go to reset its lock
state we trip over the invariant that it must still be in the wait state. 

Fix is to check if T1 was granted a lock when T2 was rolled back during
additional deadlock checking.

Approved by Jimmy Yang rb#1619.