Bug#15996571: WARNING RECOMMENDS TO USE A DEPRECATED OR NON-EXISTING OPTION
The problem was that if the startup options --log-slow-admin-statements,
--log-queries-not-using-indexes or --log-slow-slave-statements were
used without specifying --slow-query-log, a warning message referring
to the removed "--log_slow_queries" would be given instead of using the 
current name "--slow-query-log".

This patch fixes the problem by updating the warning message.
The patch also fixes similar problems with warning messages given
if --general-log-file or --slow-query-log-file are used with 
--log-output=table (and not =file).